Restructured the mod files to the common structure.
